+++ docs/conf.py
@@ -24,7 +24,7 @@ source_suffix = {
     '.md': 'markdown',
 }
 
-autodoc_mock_imports = ["plover", "PyQt5", "dulwich", "plover_cat.plover_cat_ui", "plover_cat.affix_dialog_ui", "plover_cat.caption_dialog_ui", "plover_cat.create_dialog_ui", "plover_cat.field_dialog_ui", "plover_cat.index_dialog_ui", "plover_cat.recorder_dialog_ui", "plover_cat.shortcut_dialog_ui", "plover_cat.suggest_dialog_ui", "plover_cat.test_dialog_ui"]
+autodoc_mock_imports = ["plover", "PySide6", "dulwich", "plover_cat.plover_cat_ui", "plover_cat.affix_dialog_ui", "plover_cat.caption_dialog_ui", "plover_cat.create_dialog_ui", "plover_cat.field_dialog_ui", "plover_cat.index_dialog_ui", "plover_cat.recorder_dialog_ui", "plover_cat.shortcut_dialog_ui", "plover_cat.suggest_dialog_ui", "plover_cat.test_dialog_ui"]
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output