+++ plover_cat/shortcutDialogWindow.py
@@ -1,7 +1,7 @@
-from PyQt5 import QtCore, QtGui, QtWidgets
-from PyQt5.QtCore import Qt
-from PyQt5.QtWidgets import QDialog, QMessageBox
-from PyQt5.QtGui import QStandardItemModel, QKeySequence
+from PySide6 import QtCore, QtGui, QtWidgets
+from PySide6.QtCore import Qt
+from PySide6.QtWidgets import QDialog, QMessageBox
+from PySide6.QtGui import QStandardItemModel, QKeySequence
 from plover_cat.shortcut_dialog_ui import Ui_shortcutDialog
 
 class shortcutDialogWindow(QDialog, Ui_shortcutDialog):